Rewrite it in Rust.
Added features:
  - Memory safety¹.
  - Conforming to the XDG Base Directory Specification if the
    corresponding variables are set².
  - Removing any entry for `theme` from the Ghostty configuration.

Features not added yet, but to be added in a future commit:
  - Support for reading and writing the Ghostty configuration from the
    Application Support directory on macOS.

Removed features:
  - The ability to compile and run Walors on systems which are not
    supported by Rust.  This is not actually an issue as Ghostty is not
    supported on those platforms anyway.

¹Before the rewrite, I was facing a segmentation fault due to an
out-of-bounds array access in the `line` array in `readoptions()`.  This
rewrite fixes it, as well as any other memory safety issues that may
have been lurking in the original C implementation.

²I am unsure if Pywal itself supports the XDG Base Directory
Specification or not.  A peek at the source code suggests that it in
fact has support for it, but trying to set $XDG_CACHE_HOME or
$XDG_CONFIG_HOME did not work.